// actions/dataImport/sitesData/types.ts
export enum DataImportTypes {
    SET_SITES_DATA_LIST = "DATA_SITES_IMPORT_SET_SITES_DATA_LIST",
    SET_PARSING = "DATA_SITES_IMPORT_SET_PARSING",
    SAVE_SITE_DATA = "DATA_SITES_SAVE_SITE_DATA",
    SAVE_SITE_DATA_IN_API = "DATA_SITES_SAVE_SITE_DATA_IN_API",
    DATA_SAVED = "DATA_SITES_DATA_SAVED",
    DATA_SAVED_ERROR = "DATA_SITES_DATA_SAVED_ERROR",
    SHOW_SUB_SITES = "DATA_SITES_SHOW_SUB_SITES",
    SET_SEARCHED_SITES_IN_PARENT_SITE_SELECTION = "DATA_SITES_SET_SEARCHED_SITES_IN_PARENT_SITE_SELECTION",
    RESET_SAVE_STATE = "DATA_SITES_RESET_SAVE_STATE",
    RESET_SITES_DATA = "DATA_SITES_RESET_SITES_DATA",
  }
  